Move the digest approval functions from the activity to the digest_provider.py
